hehe

+++ utils/plots.py
@@ -71,8 +71,8 @@ def plot_one_box(x, img, color=None, label=None, line_thickness=3):
 def plot_one_box_center_point(xy_center, img, color=None, line_thickness=3):
     height = int(xy_center[-1]*img.shape[0])
     width = int(xy_center[-2] * img.shape[1])
-    radius = height*width/10
-    xy_center = xy_center[:2]
+    radius = int(height*width/10)
+    xy_center = int(xy_center[:2])
     xy_center[0] = int(xy_center[0]*img.shape[1])
     xy_center[1] = int(xy_center[1] * img.shape[0])
     center_coordinates = (xy_center[0], xy_center[1])
